Crate solana_client

Modules

blockhash_query

client_error

connection_cache

mock_sender_for_cli

nonblocking

nonce_utils

Durable transaction nonce helpers.

pubsub_client

quic_client

rpc_client

Communication with a Solana node over RPC.

rpc_config

rpc_custom_error

Implementation defined RPC server errors

rpc_deprecated_config

rpc_filter

rpc_request

rpc_response

rpc_sender

A transport for RPC calls.

send_and_confirm_transactions_in_parallel

thin_client

The thin_client module is a client-side object that interfaces with a server-side TPU. Client code should use this object instead of writing messages to the network directly. The binary encoding of its messages are unstable and may change in future releases.

tpu_client

tpu_connection

transaction_executor

udp_client