Enum solana_client::connection_cache::ConnectionCache

pub enum ConnectionCache {
    Quic(Arc<BackendConnectionCache<QuicPool, QuicConnectionManager, QuicConfig>>),
    Udp(Arc<BackendConnectionCache<UdpPool, UdpConnectionManager, UdpConfig>>),
}

A thin wrapper over connection-cache/ConnectionCache to ease construction of the ConnectionCache for code dealing both with udp and quic. For the scenario only using udp or quic, use connection-cache/ConnectionCache directly.

Variants

Quic(Arc<BackendConnectionCache<QuicPool, QuicConnectionManager, QuicConfig>>)

Udp(Arc<BackendConnectionCache<UdpPool, UdpConnectionManager, UdpConfig>>)

Implementations

impl ConnectionCache

pub fn new(connection_pool_size: usize) -> Self

Create a quic connection_cache

pub fn new_with_client_options( connection_pool_size: usize, client_endpoint: Option<Endpoint>, cert_info: Option<(&Keypair, IpAddr)>, stake_info: Option<(&Arc<RwLock<StakedNodes>>, &Pubkey)> ) -> Self

Create a quic conneciton_cache with more client options

pub fn update_client_certificate( &mut self, _keypair: &Keypair, _ipaddr: IpAddr ) -> Result<(), Box<dyn Error>>

๐Deprecated since 1.15.0: This method does not do anything. Please use new_with_client_options instead to set the client certificate.

pub fn set_staked_nodes( &mut self, _staked_nodes: &Arc<RwLock<StakedNodes>>, _client_pubkey: &Pubkey )

๐Deprecated since 1.15.0: This method does not do anything. Please use new_with_client_options instead to set staked nodes information.

pub fn with_udp(connection_pool_size: usize) -> Self

pub fn use_quic(&self) -> bool

pub fn get_connection(&self, addr: &SocketAddr) -> BlockingClientConnection

pub fn get_nonblocking_connection( &self, addr: &SocketAddr ) -> NonblockingClientConnection