solana_client::nonblocking

Module nonce_utils

Durable transaction nonce helpers.

Modules

blockhash_query

Enums

Error

Functions

account_identity_ok

Perform basic checks that an account has nonce-like properties.

data_from_account

Deserialize the state data of a durable transaction nonce account.

data_from_state

Get the nonce data from its State value.

get_account

Get a nonce account from the network.

get_account_with_commitment

Get a nonce account from the network.

state_from_account

Deserialize the state of a durable transaction nonce account.