solana_client::nonce_utils::nonblocking

Function get_account_with_commitment

pub async fn get_account_with_commitment(
    rpc_client: &RpcClient,
    nonce_pubkey: &Address,
    commitment: CommitmentConfig,
) -> Result<Account, Error>

Get a nonce account from the network.

This is like RpcClient::get_account_with_commitment except:

it returns this module’s Error type,

it returns an error if the account does not exist,

it returns an error if any of the checks from account_identity_ok fail.