solana_client

Module rpc_client

Communication with a Solana node over RPC.

Software that interacts with the Solana blockchain, whether querying its state or submitting transactions, communicates with a Solana node over JSON-RPC, using the RpcClient type.

Structs

GetConfirmedSignaturesForAddress2Config

RpcClient

A client of a remote Solana node.

RpcClientConfig

Traits

SerializableMessage

Trait used to add support for versioned messages to RPC APIs while retaining backwards compatibility

SerializableTransaction

Trait used to add support for versioned transactions to RPC APIs while retaining backwards compatibility

Type Aliases

Mocks