solana_client::send_and_confirm_transactions_in_parallel

Function send_and_confirm_transactions_in_parallel_v2

pub async fn send_and_confirm_transactions_in_parallel_v2<T: Signers + ?Sized>(
    rpc_client: Arc<RpcClient>,
    tpu_client: Option<TpuClient<QuicPool, QuicConnectionManager, QuicConfig>>,
    messages: &[Message],
    signers: &T,
    config: SendAndConfirmConfigV2,
) -> Result<Vec<Option<TransactionError>>>

Sends and confirms transactions concurrently

The sending and confirmation of transactions is done in parallel tasks The method signs transactions just before sending so that blockhash does not expire.