Enum solana_client::thin_client::ThinClient

pub enum ThinClient {
    Quic(BackendThinClient<QuicPool, QuicConnectionManager, QuicConfig>),
    Udp(BackendThinClient<UdpPool, UdpConnectionManager, UdpConfig>),
}

A thin wrapper over thin-client/ThinClient to ease construction of the ThinClient for code dealing both with udp and quic. For the scenario only using udp or quic, use thin-client/ThinClient directly.

Variants

Quic(BackendThinClient<QuicPool, QuicConnectionManager, QuicConfig>)

Udp(BackendThinClient<UdpPool, UdpConnectionManager, UdpConfig>)

Implementations

impl ThinClient

pub fn new( rpc_addr: SocketAddr, tpu_addr: SocketAddr, connection_cache: Arc<ConnectionCache> ) -> Self

Create a new ThinClient that will interface with the Rpc at rpc_addr using TCP and the Tpu at tpu_addr over transactions_socket using Quic or UDP (currently hardcoded to UDP)

pub fn new_socket_with_timeout( rpc_addr: SocketAddr, tpu_addr: SocketAddr, timeout: Duration, connection_cache: Arc<ConnectionCache> ) -> Self

pub fn new_from_addrs( rpc_addrs: Vec<SocketAddr>, tpu_addrs: Vec<SocketAddr>, connection_cache: Arc<ConnectionCache> ) -> Self

pub fn rpc_client(&self) -> &RpcClient

pub fn retry_transfer_until_confirmed( &self, keypair: &Keypair, transaction: &mut Transaction, tries: usize, min_confirmed_blocks: usize ) -> TransportResult<Signature>

pub fn retry_transfer( &self, keypair: &Keypair, transaction: &mut Transaction, tries: usize ) -> TransportResult<Signature>

pub fn send_and_confirm_transaction<T: Signers + ?Sized>( &self, keypairs: &T, transaction: &mut Transaction, tries: usize, pending_confirmations: usize ) -> TransportResult<Signature>

pub fn poll_get_balance(&self, pubkey: &Pubkey) -> TransportResult<u64>

pub fn poll_get_balance_with_commitment( &self, pubkey: &Pubkey, commitment_config: CommitmentConfig ) -> TransportResult<u64>

pub fn wait_for_balance( &self, pubkey: &Pubkey, expected_balance: Option<u64> ) -> Option<u64>

pub fn get_program_accounts_with_config( &self, pubkey: &Pubkey, config: RpcProgramAccountsConfig ) -> TransportResult<Vec<(Pubkey, Account)>>

pub fn wait_for_balance_with_commitment( &self, pubkey: &Pubkey, expected_balance: Option<u64>, commitment_config: CommitmentConfig ) -> Option<u64>

pub fn poll_for_signature_with_commitment( &self, signature: &Signature, commitment_config: CommitmentConfig ) -> TransportResult<()>

pub fn get_num_blocks_since_signature_confirmation( &mut self, sig: &Signature ) -> TransportResult<usize>